Add alt-tab functionality
 - Alt-tab rotates forward; Alt-grave (the key above tab...) rotates
   backward.
 - Also fixed the bug where the taskbar wouldn't display immediately on bar
   startup.

Method: XGrab alt-tab and alt-grave. Then, for all mapped windows, set
the event mask to include key-up events (also respond to new windows
being mapped so that we can look for key-up events on those too). This
is necessary because it allows us to tell when the alt key is actually
released, so that we can hold down alt and press tab/grave repeatedly to
naturally rotate the windows.